# openapi.yaml
openapi: 3.0.3
info:
  title: Neuro Resume API
  description: |
    API для приложения создания резюме с помощью AI-интервью.
    Приложение проводит интерактивное интервью с пользователем,
    собирает информацию о его опыте и навыках, а затем генерирует
    профессиональное резюме.
  version: 1.0.0
  contact:
    name: Neuro Resume Team
    email: support@neuroresume.com
servers:
  - url: https://api.neuroresume.com/v1
    description: Production server
  - url: https://staging-api.neuroresume.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
paths:
...
  /resumes:
    get:
      tags:
        - resume
      summary: Получить список резюме
      description: Все резюме текущего пользователя
      operationId: getResumes
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Список резюме
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resume'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /resumes/{resumeId}:
    get:
      tags:
        - resume
      summary: Получить резюме
      description: Детальная информация о резюме
      operationId: getResume
      security:
        - bearerAuth: []
      parameters:
        - name: resumeId
          in: path
          required: true
          description: ID резюме
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Резюме успешно получено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - resume
      summary: Обновить резюме
      description: Редактирование данных резюме
      operationId: updateResume
      security:
        - bearerAuth: []
      parameters:
        - name: resumeId
          in: path
          required: true
          description: ID резюме
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeData'
      responses:
        '200':
          description: Резюме успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - resume
      summary: Удалить резюме
      description: Удаление резюме
      operationId: deleteResume
      security:
        - bearerAuth: []
      parameters:
        - name: resumeId
          in: path
          required: true
          description: ID резюме
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Резюме успешно удалено
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /resumes/{resumeId}/download:
    get:
      tags:
        - resume
      summary: Скачать резюме
      description: Получить резюме в выбранном формате
      operationId: downloadResume
      security:
        - bearerAuth: []
      parameters:
        - name: resumeId
          in: path
          required: true
          description: ID резюме
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          required: true
          description: Формат файла
          schema:
            type: string
            enum: [pdf, docx, txt]
            example: pdf
      responses:
        '200':
          description: Файл резюме
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /resumes/{resumeId}/regenerate:
    post:
      tags:
        - resume
      summary: Перегенерировать резюме
      description: Создать новую версию резюме на основе тех же данных интервью
      operationId: regenerateResume
      security:
        - bearerAuth: []
      parameters:
        - name: resumeId
          in: path
          required: true
          description: ID резюме
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                template:
                  type: string
                  description: Шаблон для генерации
                  enum: [modern, classic, minimal, creative]
                  example: modern
                language:
                  type: string
                  enum: [ru, en]
                  example: ru
      responses:
        '200':
          description: Резюме перегенерировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный при авторизации
  schemas:
...
    Resume:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 321e6543-e89b-12d3-a456-426614174333
        userId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
          description: ID интервью-сессии, из которой создано резюме
        title:
          type: string
          example: Senior Frontend Developer
        template:
          type: string
          enum: [modern, classic, minimal, creative]
          example: modern
        language:
          type: string
          enum: [ru, en]
          example: ru
        data:
          $ref: '#/components/schemas/ResumeData'
        version:
          type: integer
          description: Версия резюме (для отслеживания изменений)
          example: 3
        createdAt:
          type: string
          format: date-time
          example: 2025-10-12T11:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-10-12T12:30:00Z
    ResumeData:
      type: object
      properties:
        personalInfo:
          type: object
          properties:
            firstName:
              type: string
              example: Иван
            lastName:
              type: string
              example: Петров
            email:
              type: string
              format: email
              example: ivan@example.com
            phone:
              type: string
              example: +7 (999) 123-45-67
            location:
              type: string
              example: Москва, Россия
            links:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [linkedin, github, telegram, website, portfolio]
                  url:
                    type: string
                    format: uri
              example:
                - type: github
                  url: https://github.com/johndoe
                - type: linkedin
                  url: https://linkedin.com/in/johndoe
        summary:
          type: string
          description: Краткое описание профессионала
          example: Опытный Frontend разработчик с 5+ годами опыта...
        workExperience:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
                example: Tech Corp
              position:
                type: string
                example: Senior Frontend Developer
              location:
                type: string
                example: Москва
              startDate:
                type: string
                format: date
                example: 2020-03
              endDate:
                type: string
                format: date
                nullable: true
                example: null
              current:
                type: boolean
                example: true
              description:
                type: string
                example: Разработка и поддержка веб-приложений...
              achievements:
                type: array
                items:
                  type: string
                example:
                  - Увеличил производительность на 40%
                  - Внедрил CI/CD pipeline
        education:
          type: array
          items:
            type: object
            properties:
              institution:
                type: string
                example: МГУ
              degree:
                type: string
                example: Бакалавр
              field:
                type: string
                example: Информатика
              startDate:
                type: string
                format: date
                example: 2015-09
              endDate:
                type: string
                format: date
                example: 2019-06
              gpa:
                type: number
                format: float
                nullable: true
                example: 4.5
        skills:
          type: object
          properties:
            technical:
              type: array
              items:
                type: string
              example: [React, TypeScript, Node.js, PostgreSQL]
            soft:
              type: array
              items:
                type: string
              example: [Командная работа, Лидерство, Коммуникация]
            languages:
              type: array
              items:
                type: object
                properties:
                  language:
                    type: string
                    example: Английский
                  level:
                    type: string
                    enum: [A1, A2, B1, B2, C1, C2, native]
                    example: C1
        certifications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: AWS Certified Developer
              issuer:
                type: string
                example: Amazon Web Services
              date:
                type: string
                format: date
                example: 2023-05
              expirationDate:
                type: string
                format: date
                nullable: true
              credentialId:
                type: string
                nullable: true
        projects:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: E-commerce Platform
              description:
                type: string
              technologies:
                type: array
                items:
                  type: string
              url:
                type: string
                format: uri
                nullable: true
              startDate:
                type: string
                format: date
              endDate:
                type: string
                format: date
                nullable: true
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Общее количество записей
          example: 150
        limit:
          type: integer
          description: Количество записей на странице
          example: 20
        offset:
          type: integer
          description: Текущее смещение
          example: 0
        hasMore:
          type: boolean
          description: Есть ли еще записи
          example: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Invalid email format
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
                - field: password
                  message: Password must be at least 6 characters